# app/Http/Resources/Poll.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Resources\Json\JsonResource;

class Poll extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return array
     */
    public function toArray($request)
    {
        return [
            'title' => $this->title,
            'last_modified' => (string) $this->updated_at
        ];
    }

    // public function with($request){
    //     return ['success' => true];
    // }
}